style(bench): fix Linux-cfg-only clippy lints invisible to macOS dev loop
Round-5's cold local clippy ran on macOS, so code behind
`#[cfg(target_os = "linux")]` was never compiled — CI (ubuntu) caught:

- eval/memory.rs sample_rss: `stat.rss as u64` — procfs already types
  `rss` as u64; drop the cast (clippy::unnecessary_cast).
- helios_p50 + consolidator_tick benches /proc/cpuinfo probe: collapse
  the nested `if let Some(rest)` / `if let Some(idx)` into a single
  let-chain (clippy::collapsible_if) — preemptive: same shape as the
  site CI flagged, sitting behind the same Linux cfg in bench targets
  that compile right after the failed lib.

Audited every remaining `target_os = "linux"` block in the workspace
(grep): no other Linux-only lint candidates.

footer: PR #11 round 6
author: Tin Dang

+++ crates/lunaris-bench/src/eval/memory.rs
@@ -119,7 +119,9 @@ fn sample_rss() -> Option<u64> {
     let proc = procfs::process::Process::myself().ok()?;
     let stat = proc.stat().ok()?;
     let page_size = procfs::page_size();
-    Some((stat.rss as u64) * page_size)
+    // `stat.rss` is already u64 in the pinned procfs — no cast
+    // (clippy::unnecessary_cast under the Linux cfg).
+    Some(stat.rss * page_size)
 }
 
 /// macOS RSS sampler — shells out to `ps -o rss= -p <pid>` (returns RSS in
